+++ app/router/router.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class Empty(BaseModel):
+    pass
