+++ app/router/user.py
@@ -1,0 +1,55 @@
+from fastapi import APIRouter, Depends, HTTPException
+from fastapi.security.http import HTTPAuthorizationCredentials, HTTPBearer
+from pydantic import BaseModel
+
+from app.router.router import Empty
+
+
+from app import model
+
+
+router = APIRouter()
+
+
+class UserCreateRequest(BaseModel):
+    user_name: str
+    leader_card_id: int
+
+
+class UserCreateResponse(BaseModel):
+    user_token: str
+
+
+@router.post("/create", response_model=UserCreateResponse)
+def user_create(req: UserCreateRequest):
+    """新規ユーザー作成"""
+    token = model.create_user(req.user_name, req.leader_card_id)
+    return UserCreateResponse(user_token=token)
+
+
+bearer = HTTPBearer()
+
+
+def get_auth_token(cred: HTTPAuthorizationCredentials = Depends(bearer)) -> str:
+    assert cred is not None
+    if not cred.credentials:
+        raise HTTPException(status_code=401, detail="invalid credential")
+    return cred.credentials
+
+
+@router.get("/me", response_model=model.SafeUser)
+def user_me(token: str = Depends(get_auth_token)):
+    user = model.get_user_by_token(token)
+    if user is None:
+        raise HTTPException(status_code=404)
+    # print(f"user_me({token=}, {user=})")
+    return user
+
+
+# response_model で Type Hintをつけるのは、空のDictを返すときに便利
+@router.post("/update", response_model=Empty)
+def update(req: UserCreateRequest, token: str = Depends(get_auth_token)):
+    """Update user attributes"""
+    # print(req)
+    model.update_user(token, req.user_name, req.leader_card_id)
+    return {}
